removes unused repos
