+++ frontend/src/Services/ReduxStore/Reducers/TaskResult/index.tsx
@@ -1,7 +1,7 @@
 import { AnyAction, createEntityAdapter, createSlice, ThunkAction } from '@reduxjs/toolkit';
-import TaskResultRepo, { TaskResult } from '../../../../Repos/TaskResultRepo';
+import { useOperationMethod } from 'react-openapi-client/useOperationMethod';
+import { TaskResult } from '../../../../Components/Shared/Interfaces/TaskResult';
 import { RootState } from '../../Store';
-
 
 
 const taskResultsAdapter = createEntityAdapter<TaskResult>({
@@ -34,16 +34,20 @@ export default taskResultsSlice.reducer;
 // TODO: use createAsyncThunk/AsyncThunkAction instead
 export const fetchTaskResults = (): ThunkAction<void, RootState, unknown, AnyAction> => async dispatch => {
   try {
-    const taskResultRepo = new TaskResultRepo();
-    const response = await taskResultRepo.findAll(
-      { page: 1, pageSize: 10, filters: { 'filter[taskName.icontains]': 'build_ogc_service' } }
-    );
+    const [listTaskResult, { response: listTaskResultResponse }] = useOperationMethod('listTaskResult');
+    listTaskResult([
+      { name: 'page[number]', value: 1, in: 'query' },
+      { name: 'page[size]', value: 10, in: 'query' },
+      { name: 'filter[taskName.icontains]', value: 'build_ogc_service', in: 'query' }
+    ]);
+
+
     // TODO: move error/nulltype handling to Repo
-    if (response.data && response.data.data) {
+    if (listTaskResultResponse.data && listTaskResultResponse.data.data) {
       dispatch(
         {
           type: 'taskResults/set',
-          payload: response.data.data
+          payload: listTaskResultResponse.data.data
         }
       );
     }
